dev-perl/Dist-Zilla: Add to tree at version 5.39.0
required by app-admin/rex for doing VCS installs.

NOTE: This perl module is an "authorside" module and as such it is
not TYPICALLY used by anything for installation, as the point of
Dist::Zilla is to create simplifed distributions that don't need
a lot of install-time magic.

It is subsequently *not* used on typical CPAN sources, and should
only be used against VCS sources to create an analog of a published
CPAN distrubition as part of the VCS build workflow.

Previously maintained in ::perl-experimental by Torsten Veller,
Michele Beltrame and myself since May 2009

Package-Manager: portage-2.2.20.1